Update ranges to be inclusive of the end number
Previously this would not detect a date such as "2020-01-31" or "2019-12-01" because "12" and "31" were not valid values for months or years. Looking through this code, this same issue appears to apply to the other ranges as well so they have all been adjusted.

This also fixes a bug where if a date like "2020-01-01" was used, it would properly detect it as a 2 digit date but it would result in "%Y-%m-%m" instead of "%Y-%m-%d" because "01" would be replaced with "%m" twice. You can use the third argument to `str.replace` to tell Python to only replace 1 instance, which solves this issue.

This also removes a duplicate check for `SHORT_MONTHS` which was already being done above.